refactored: to keep methods small
